bugbounty-scanner/requirements.txt
# ============================================================
# AI Bug Bounty Scanner - Python Dependencies
# ============================================================

# Web Framework & API
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# HTTP & Async
aiohttp==3.9.1
httpx==0.25.2
requests==2.31.0
websockets==12.0

# HTML Parsing & Crawling
beautifulsoup4==4.12.2
lxml==4.9.3
html5lib==1.1
playwright==1.40.0

# DNS & Network Recon
dnspython==2.4.2
python-whois==0.8.0
scapy==2.5.0
python-nmap==0.7.1
ipwhois==1.2.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
asyncpg==0.29.0
psycopg2-binary==2.9.9

# AI / ML
scikit-learn==1.3.2
numpy==1.26.2
pandas==2.1.3
joblib==1.3.2

# Payload Analysis
regex==2023.10.3
urllib3==2.1.0
yarl==1.9.4

# Reporting
jinja2==3.1.2
weasyprint==60.1
markdown==3.5.1
pdfkit==1.0.0

# CLI
click==8.1.7
rich==13.7.0
typer==0.9.0

# Security & Crypto
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==41.0.7

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
loguru==0.7.2
tqdm==4.66.1
colorama==0.4.6
tabulate==0.9.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Rate limiting & Retry
tenacity==8.2.3
ratelimit==2.2.1
